The engine is the loop.
Move most code of the Engine class to the new Simulation class.
Get rid of the 'loop' module. We will have only one loop any way.

I want to have a clear view over the tricky code of the engine which
keeps track of the time in the simulation and the refresh rate.